+++ hittable.h
@@ -9,7 +9,7 @@
 class Renderable
 {
 public:
-    virtual bool hit(Math::Ray ray, float t0, float t1, std::shared_ptr<Util::HitRecord> & hitRecord) const = 0;
+    virtual bool hit(Math::Ray ray, double t0, double t1, std::shared_ptr<Util::HitRecord> & hitRecord) const = 0;
     virtual Util::Color computeColor(
         const std::vector<std::unique_ptr<LightSource>> & lightSources,
         Math::Ray viewRay,
